regression/requirements.txt
boto3 == 1.2.3
matplotlib == 1.3.1
munkres == 1.0.7
pandas == 0.16.2
pylearn2.egg == info
scipy == 0.16.1
Theano == 0.7.0
Pillow == 3.2.0
numpy == 1.11.1
scikit_learn == 0.17.1
